feat(charsheet): Add weight limit helpers
Added method to reject items during drag/drop if the item would
place the character over weight limit. Added friendly notice that
the chracter is Encumbered once dragging an item would place the
chracter in that state.  Also added a tooltip to the "Total Weight"
display listing the character's Unencumbered and Encumbered maximums.